requirements.txt
python-dotenv
selenium
beautifulsoup4
discord